refactor: reconstruct package for gamerule
