+++ src/main/java/racing/domain/gamerule/GameRule.java
@@ -1,4 +1,4 @@
-package racing.domain;
+package racing.domain.gamerule;
 
 import racing.common.exception.InvalidInputFormatException;
 import racing.domain.player.RacingPlayer;
